Ease impl and readability by keeping interfaces together
