--- workerpool/config.go
@@ -1,8 +1,0 @@
-package workerpool
-
-// Config is the initial configuration of the workerpool
-type Config struct {
-	InitialNumProducers uint
-	InitialNumConsumers uint
-	WorkQueueDepth      uint
-}
