--- workerpool/consumer.go
@@ -1,7 +1,0 @@
-package workerpool
-
-import "context"
-
-type Consumer interface {
-	Run(context.Context, ThreadID) error
-}
