--- workerpool/consumer_factory.go
@@ -1,6 +1,0 @@
-package workerpool
-
-type ConsumerFactory interface {
-	New(SubmissionQueue) (Consumer, error)
-	Finished(ThreadID, Consumer)
-}
