+++ workerpool/example/main.go
@@ -12,6 +12,18 @@ import (
 	"github.com/sean-/sysexits"
 )
 
+// Interface is the workerpool interface that is exported
+type Interface interface {
+	// ShutdownCtx is the shutdown context, used to poll for the status of shutdown
+	ShutdownCtx() context.Context
+
+	// InitiateShutdown cancels the context used to signal its time to shut down
+	InitiateShutdown() (bool, error)
+
+	// Reload is a configuration reload handler (e.g. for use with SIGHUP)
+	Reload()
+}
+
 func main() {
 	os.Exit(realMain())
 }
@@ -81,7 +93,7 @@ func realMain() int {
 	return sysexits.OK
 }
 
-func runSignalHandler(wp workerpool.Interface) error {
+func runSignalHandler(wp Interface) error {
 	sigCh := make(chan os.Signal, 5)
 
 	signal.Notify(sigCh,
