--- workerpool/factories.go
@@ -1,10 +1,0 @@
-package workerpool
-
-// ThreadID is a generic identifier for a given thread
-type ThreadID int
-
-// Factories is a container for different types of factories
-type Factories struct {
-	ProducerFactory ProducerFactory
-	ConsumerFactory ConsumerFactory
-}
