--- workerpool/producer.go
@@ -1,7 +1,0 @@
-package workerpool
-
-import "context"
-
-type Producer interface {
-	Run(context.Context, ThreadID) error
-}
