--- workerpool/producer_factory.go
@@ -1,6 +1,0 @@
-package workerpool
-
-type ProducerFactory interface {
-	New(SubmissionQueue) (Producer, error)
-	Finished(ThreadID, Producer)
-}
