--- workerpool/submissionqueue.go
@@ -1,3 +1,0 @@
-package workerpool
-
-type SubmissionQueue chan Task
