--- workerpool/task.go
@@ -1,4 +1,0 @@
-package workerpool
-
-type Task interface {
-}
